requirements_distributed.txt
# Enhanced requirements for distributed training
# Base requirements
annotated-types==0.6.0
blis==0.7.11
brotlipy==0.7.0
catalogue==2.0.10
certifi 
cffi 
charset-normalizer 
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.3
contourpy==1.1.1
cryptography 
cycler==0.12.1
cymem==2.0.8
Farama-Notifications==0.0.4
fonttools==4.43.1
gym
gym-notices
gymnasium
hashids==1.3.1
jericho==3.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mementos==1.3.1
more-itertools==10.1.0
murmurhash==1.0.10
networkx==3.2.1
numpy
packaging==23.2
Pillow
pluggy
preshed==3.0.9
prompt-toolkit==3.0.39
pycosat 
pycparser 
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL 
pyparsing==3.1.1
PySockets 
python-dateutil==2.8.2
requests 
ruamel.yaml 
ruamel.yaml.clib 
six
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
TatSu
thinc==8.2.1
toolz 
tqdm 
typer==0.9.0
typing_extensions==4.8.0
urllib3 
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
zstandard 
hydra-core
gradio
beautifulsoup4
sentencepiece

# Core ML/DL packages for distributed training
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.37.2
tokenizers>=0.15.0

# Distributed training frameworks
accelerate>=0.25.0
deepspeed>=0.12.0
fairscale>=0.4.13

# Memory optimization
bitsandbytes>=0.41.0
flash-attn>=2.4.0

# PEFT for parameter-efficient fine-tuning
peft>=0.7.0
loralib>=0.1.1

# Model serving and inference optimization
vllm>=0.2.5
text-generation-inference>=1.4.0

# Communication and networking
nccl-cuda>=2.18.0
mpi4py>=3.1.4

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
psutil>=5.9.0
GPUtil>=1.4.0
pynvml>=11.5.0

# API and cloud integration
openai>=1.3.0
huggingface-hub>=0.19.0
datasets>=2.15.0

# Development and debugging
ipython>=8.17.0
jupyter>=1.0.0
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0

# Configuration management
omegaconf>=2.3.0
pyyaml>=6.0.1

# Process management for distributed training
supervisor>=4.2.5
fabric>=3.2.0

# Performance profiling
py-spy>=0.3.14
line_profiler>=4.1.1
memory_profiler>=0.61.0

# Visualization
seaborn>=0.12.2
plotly>=5.17.0 